requirements.txt
Flask==3.0.0
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
python-docx==1.1.0
Werkzeug==3.0.1

